Fix FATEsanity in Client
