+++ src/hooks/duties.csv
@@ -7,7 +7,7 @@ Halatali,Dungeon (ARR),23,23,Eastern Thanalan,Solo,Normal,
 The Thousand Maws of Toto-Rak,Dungeon (ARR),27,27,South Shroud,Solo,Normal,
 Haukke Manor,Dungeon (ARR),31,31,Central Shroud,Solo,Normal,
 Brayflox's Longstop,Dungeon (ARR),34,34,Eastern La Noscea,Solo,Normal,
-The Sunken Temple of Qarn ,Dungeon (ARR),37,37,Southern Thanalan,Solo,Normal,
+The Sunken Temple of Qarn,Dungeon (ARR),37,37,Southern Thanalan,Solo,Normal,
 Cutter's Cry,Dungeon (ARR),40,40,Central Thanalan,Solo,Normal,
 The Stone Vigil,Dungeon (ARR),43,43,Coerthas Central Highlands,Solo,Normal,
 Dzemael Darkhold,Dungeon (ARR),46,46,Coerthas Central Highlands,Solo,Normal,
